auth_service/requirements.txt
Django>=4.2
djangorestframework
mysqlclient
djangorestframework-simplejwt
gunicorn